requirements.txt
python-dotenv
huggingface_hub
temporalio
numpy 
torch
neuraloperator